+++ extensions/iceworks-material-helper/src/services/index.ts
@@ -1,7 +1,6 @@
 import * as material from '@iceworks/material-service';
-import * as page from '@iceworks/page-service';
 import * as common from '@iceworks/common-service';
 import * as block from '@iceworks/block-service';
 import * as component from '@iceworks/component-service';
 
-export default { material, page, common, block, component };
+export default { material, common, block, component };