+++ packages/component-service/src/index.ts
@@ -9,6 +9,7 @@ import {
   getDataFromSettingJson,
   getIceworksTerminal,
   checkPathExists,
+  showTextDocument,
 } from '@iceworks/common-service';
 import {
   jsxFileExtnames,
@@ -163,13 +164,22 @@ export async function generateComponentCode(
   const schema: IBasicSchema = { version, componentsMap, componentsTree: [componentsTree], i18n, utils };
   try {
     await generateCode(componentName, schema);
-    vscode.window.showInformationMessage(
-      i18nService.format('package.component-service.index.createComponentSuccess', {
-        componentPath: path.join(componentsPath, componentName),
-      })
-    );
   } catch (e) {
     vscode.window.showErrorMessage(e.message);
+    throw e;
+  }
+
+  const componentPath = path.join(componentsPath, componentName);
+  const openFileAction = i18nService.format('package.component-service.index.openFile');
+  const selectedAction = await vscode.window.showInformationMessage(
+    i18nService.format('package.component-service.index.createComponentSuccess', {
+      componentPath,
+    }),
+    openFileAction
+  );
+
+  if (selectedAction === openFileAction) {
+    showTextDocument(componentPath);
   }
 }