+++ extensions/iceworks-ui-builder/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.11
 
 - feat: show open file action after create component
+- refactor: router logic
 
 ## 0.1.10