+++ extensions/iceworks-ui-builder/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.1.11
+
+- feat: show open file action after create component
+- refactor: router logic
+
+## 0.1.10
+
+- fix: generate block code error
+
 ## 0.1.9
 
 - fix: build error